requirements.txt
crewai>=0.40.0
crewai-tools>=0.4.0
langchain_community>=0.2.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
faiss-cpu>=1.8.0
langchain-google-genai>=1.0.0
langchain>=0.2.0
langchain-core>=0.2.0
litellm>=1.0.0
sentence-transformers>=2.2.0
